Refactor registration and access
Registration was broken out into individual components instead of using
the consts that were originally declared.  IMO it's a bit cleaner and
lets us handle differences between the types easier in the future
(though they are pretty similar now).  Error handling for unfound values
was changed slightly.  If a metric is missing, we increment specific
counters for the type and in the case of timers we send back a special
stub so the caller can still defer the observe.