Refactor for summary options support (#10)
Added support to override some of the summary options that prometheus
allows.  Because there were several options that we could support.  In
order to maintain consistency, histogram was updated to also use options
as well.  The only thing that I don't like (and this existed before
refactor), is the requirement to pass in the options each time you call
one of the metric functions even though it is used during the
registration of the new metric.  It's not extremely horrible, but
something that is just a bit annoying.  The actual metrics packages were
moved up into the apex package due to circular dependencies that
developed.