feat(server): add explict shutdown
Back away from using letting the user provide a base context
and use an explicit shutdown method to gracefully shutdown.
The problem with allowing the base context is if that context
is canceled via signal or some other method, the http server
will throw back 504s for all new and existing requests which
can result in bad or unexpected behavior to the client.

* [x] Add a grace period that is honored before the shutdown to allow the scrapers time to get the last metrics before shutting down.
* [x] Remove the base context.
* [x] Add the logging implementation so we can get some useful logs.  By default a "nil" logger is created that ignores all logs.
* [x] Update examples to show the explicit stop.