feat(context): context functions
Provides context transport and extraction functions. This is still early
and I want a chance to test it out in a couple places before it is
"officially" added to the documentation.

As an aside, when working through the context components I thought that
it would be useful to have an implementation that could just return a
"discard" implementation of the metrics instead of returning nil.  I'm
considering a refactor to pull out Metrics into an interface so I can
add that Discard implementation and maybe actually extend this to handle
more than just prometheus. For now, this should be fine.